refactor: relpace float to double

+++ src/base/camera.h
@@ -6,5 +6,5 @@ class Camera {
     public:
         __device__ virtual ~Camera() {}
 
-        __device__ virtual Ray get_ray(float s, float t, curandState *local_rand_state) const = 0;
+        __device__ virtual Ray get_ray(double s, double t, curandState *local_rand_state) const = 0;
 };
