+++ src/base/world.h
@@ -9,10 +9,10 @@ class World {
 
         __device__ World(Shape **_list, int n) : list(_list), size(n) {}
 
-        __device__ bool intersect(Intersection &intersection, const Ray &ray, float t_min,
-                                  float t_max) const {
+        __device__ bool intersect(Intersection &intersection, const Ray &ray, double t_min,
+                                  double t_max) const {
             bool intersected = false;
-            float closest_so_far = t_max;
+            double closest_so_far = t_max;
             Intersection temp_intersection;
             for (int idx = 0; idx < size; idx++) {
                 if (list[idx]->intersect(temp_intersection, ray, t_min, closest_so_far)) {
