+++ src/cameras/perspective_camera.h
@@ -15,13 +15,13 @@ class PerspectiveCamera : public Camera {
     public:
         ~PerspectiveCamera() override = default;
 
-        __device__ PerspectiveCamera(Point look_from, Point look_at, Vector3 up, float vfov, float aspect,
-                                     float aperture, float focus_dist) {
+        __device__ PerspectiveCamera(Point look_from, Point look_at, Vector3 up, double vfov, double aspect,
+                                     double aperture, double focus_dist) {
             auto PI = acos(-1.0);
             lens_radius = aperture / 2.0f;
-            float theta = vfov * PI / 180.0f;
-            float half_height = tan(theta / 2.0f);
-            float half_width = aspect * half_height;
+            double theta = vfov * PI / 180.0f;
+            double half_height = tan(theta / 2.0f);
+            double half_width = aspect * half_height;
             origin = look_from;
             w = (look_from - look_at).normalize();
             u = cross(up, w).normalize();
@@ -32,7 +32,7 @@ class PerspectiveCamera : public Camera {
             vertical = 2.0f * half_height * focus_dist * v;
         }
 
-        __device__ Ray get_ray(float s, float t, curandState *local_rand_state) const override {
+        __device__ Ray get_ray(double s, double t, curandState *local_rand_state) const override {
             Vector3 rd = lens_radius * random_in_unit_disk(local_rand_state);
             Vector3 offset = u * rd.x + v * rd.y;
             return Ray(origin + offset, lower_left_corner + s * horizontal + t * vertical - origin - offset);
@@ -43,5 +43,5 @@ class PerspectiveCamera : public Camera {
         Vector3 horizontal;
         Vector3 vertical;
         Vector3 u, v, w;
-        float lens_radius;
+        double lens_radius;
 };
