+++ src/shapes/sphere.h
@@ -5,28 +5,28 @@
 class Sphere : public Shape {
     private:
         Point center;
-        float radius;
+        double radius;
         const Material *material_ptr;
 
     public:
         ~Sphere() override = default;
 
-        __device__ Sphere(const Point &_center, float _radius, const Material *_material_ptr)
+        __device__ Sphere(const Point &_center, double _radius, const Material *_material_ptr)
             : center(_center), radius(_radius), material_ptr(_material_ptr) {}
 
-        __device__ bool intersect(Intersection &intersection, const Ray &ray, float t_min,
-                                  float t_max) const override {
+        __device__ bool intersect(Intersection &intersection, const Ray &ray, double t_min,
+                                  double t_max) const override {
             Vector3 oc = ray.o - center;
-            float a = dot(ray.d, ray.d);
-            float b = dot(oc, ray.d);
-            float c = dot(oc, oc) - radius * radius;
-            float discriminant = b * b - a * c;
+            double a = dot(ray.d, ray.d);
+            double b = dot(oc, ray.d);
+            double c = dot(oc, oc) - radius * radius;
+            double discriminant = b * b - a * c;
 
             if (discriminant < 0.0) {
                 return false;
             }
 
-            float temp = (-b - sqrt(discriminant)) / a;
+            double temp = (-b - sqrt(discriminant)) / a;
             if (temp < t_max && temp > t_min) {
                 intersection.t = temp;
                 intersection.p = ray.at(intersection.t);
